+++ include/glk/pointcloud_buffer_pcl.hpp
@@ -11,7 +11,7 @@ namespace glk {
 template<typename PointT>
 std::shared_ptr<PointCloudBuffer> create_point_cloud_buffer(const pcl::PointCloud<PointT>& cloud) {
   std::vector<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f>> points(cloud.size());
-  for(int i = 0; i < cloud.size(); i++) {
+  for(size_t i = 0; i < cloud.size(); i++) {
     points[i] = cloud[i].getVector3fMap();
   }
   return std::make_shared<PointCloudBuffer>(points[0].data(), sizeof(Eigen::Vector3f), cloud.size());
@@ -26,7 +26,7 @@ template<>
 inline std::shared_ptr<PointCloudBuffer> create_point_cloud_buffer(const pcl::PointCloud<pcl::PointNormal>& cloud) {
   std::vector<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f>> points(cloud.size());
   std::vector<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f>> normals(cloud.size());
-  for(int i = 0; i < cloud.size(); i++) {
+  for(size_t i = 0; i < cloud.size(); i++) {
     points[i] = cloud[i].getVector3fMap();
     normals[i] = cloud[i].getNormalVector3fMap();
   }
@@ -42,7 +42,7 @@ std::shared_ptr<PointCloudBuffer> create_colored_point_cloud_buffer(const pcl::P
   std::vector<Eigen::Vector3f, Eigen::aligned_allocator<Eigen::Vector3f>> points(cloud.size());
   std::vector<Eigen::Vector4f, Eigen::aligned_allocator<Eigen::Vector4f>> colors(cloud.size());
 
-  for(int i = 0; i < cloud.size(); i++) {
+  for(size_t i = 0; i < cloud.size(); i++) {
     points[i] = cloud[i].getVector3fMap();
     colors[i] = Eigen::Vector4f(cloud[i].r, cloud[i].g, cloud[i].b, cloud[i].a) / 255.0f;
   }
